# docs/swagger.yaml
definitions:
  api.ErrorResponse:
    description: Generic error response
    properties:
      error:
        description: Contains the error message
        type: string
    type: object
  api.GetMeResponse:
    description: Contains the GUID for the authenticated user
    properties:
      guid:
        example: 12345678-1234-1234-1234-123456789012
        type: string
    type: object
  api.LoginRequest:
    properties:
      guid:
        description: GUID for the user that is logging in
        example: 12345678-1234-1234-1234-123456789012
        type: string
    required:
    - guid
    type: object
  api.RefreshRequest:
    properties:
      refresh_token:
        example: base64-encoded-token
        type: string
    required:
    - refresh_token
    type: object
  api.TokenPair:
    properties:
      access_token:
        description: AccessToken is a JWT token that can be used to access the API
        type: string
      refresh_token:
        description: |-
          RefreshToken is a randomly generated base64 string that can be used to refresh the access token
          It is valid for 30 days
          Refresh token can only be used to refresh a single access token it was issued with.
          After refreshing, the refresh token is no longer valid and cannot be used again.
        type: string
    type: object
info:
  contact: {}
paths:
  /login:
    post:
      consumes:
      - application/json
      parameters:
      - description: login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.TokenPair'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Generate a token pair from guid
  /logout:
    delete:
      description: Deletes the auth for the authenticated user
      responses:
        "204":
          description: Successfully logged out
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Logout the authenticated user
  /me:
    get:
      description: Returns the GUID for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GetMeResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get the GUID for the authenticated user
  /refresh:
    put:
      consumes:
      - application/json
      description: Refresh the access token for the authenticated user
      parameters:
      - description: refresh request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.TokenPair'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Refresh the access token for the authenticated user
securityDefinitions:
  BearerAuth:
    description: Authorization header using the Bearer scheme
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"